7월 8일 배포 2차

+++ trade-service/src/main/java/com/moni/trade/asset/application/calculator/AssetCalculator.java
@@ -121,7 +121,7 @@ public class AssetCalculator {
             throw new CustomException(AssetErrorCode.STOCK_PRICE_NOT_FOUND);
         }
 
-        BigDecimal quantity = BigDecimal.valueOf(holding.quantity());
+        BigDecimal quantity = holding.quantity();
         BigDecimal purchaseAmount = holding.totalPurchaseAmount();
         BigDecimal evaluationAmount = price.currentPrice().multiply(quantity);
         BigDecimal profitLoss = evaluationAmount.subtract(purchaseAmount);
